+++ src/components/index.ts
@@ -1,2 +1,2 @@
 export { default as Header } from "./Header/Header";
-export { default as ProjectModal } from "./Project/ProjectModal/ProjectModal";
+export { default as ProjectModal } from "./Project/ProjectModal/Project";
